modify io_pymc for v3&4 compatibility
